[examples] Update GameOfLife to be console only.
 - The HTTP library is no longer maintained, so for now just change this example
   to print to the console.
