--- examples/GameOfLife/Static/index.html
@@ -1,15 +1,0 @@
-<!-- This source file is part of the Swift.org open source project
-
- Copyright 2017 Apple Inc. and the Swift project authors
- Licensed under Apache License v2.0 with Runtime Library Exception
-
- See http://swift.org/LICENSE.txt for license information
- See http://swift.org/CONTRIBUTORS.txt for Swift project authors -->
-
-<html lang="en">
-  <head><title>Game of Life</title></head>
-  <body onload="init()">
-    <svg id="lifeBoard" width="160" height="160"/>
-    <script type="text/javascript" src="Static/PlayLife.js"></script>
-  </body>
-</html>
